improve error handling
To unify the output, errors during the initial parsing now also go to
stderr.

The previously used Ap.parseOnly discards any leftover contents of the
file if that was not parsed. To show these errors to the user, the code
was changed to use Ap.parse instead. The leftover contents is directly
feed to stderr (a possible improvement could be to only display enough
context so that the user can find the location in the file). The parsing
still returns everything which could be parsed, so that the run can
continue.
The adding of missing curly brackets was restructured so that it also
handles cases in which the beginning of the file was parsed successfully
(the result is not empty in that case).