add handleModifier for eu4 modifiers
it accepts a localization key and a value transformation function as
parameters. The icon is assumed to be just the localized name.
Using this function avoids having one message for each modifier and
keeps everything about a modifier in the same place. which makes it
easier to find and fix it and adding new modifiers just needs one new
line.

for now it is only used for the new tolerance_of_heretics_capacity and
tolerance_of_heathens_capacity which is related and
yearly_doom_reduction which was missing an icon.