add missing eu4 modifiers
some of them are currently only usable in ages, tech or flagships,
but this is supposed to get changed in the next patch.

They all use either handleModifier, handleModifierWithIcon or rhsAlwaysYes
as handler. I added a new handlersForModifiers map for them to which I
also moved the three previous modifiers which used handleModifier