different handling for modifiers
EU4Bonus scope now prefers handlers from handlersForModifiers
which are not used for any other scope anymore. Eventually all modifier
handlers should be moved to handlersForModifiers and the fallback to use
other handlers should be removed. As a first step, some of the numericIconBonus
lines were split up into their modifier and the non-modifier message.
This allows the modifiers to use a better icon.
In the process it was discovered that devotion lacked a message for its
use as a trigger which was added as well.